+++ openhands/integrations/github/service/prs.py
@@ -1,5 +1,5 @@
 from openhands.core.logger import openhands_logger as logger
-from openhands.integrations.github.service.base import GitHubHTTPClient
+from openhands.integrations.github.github_http_client import GitHubHTTPClient
 from openhands.integrations.service_types import RequestMethod
