+++ openhands/integrations/github/service/resolver.py
@@ -2,12 +2,12 @@ from datetime import datetime
 from typing import Any
 
 from openhands.core.logger import openhands_logger as logger
+from openhands.integrations.github.github_http_client import GitHubHTTPClient
 from openhands.integrations.github.queries import (
     get_review_threads_graphql_query,
     get_thread_comments_graphql_query,
     get_thread_from_comment_graphql_query,
 )
-from openhands.integrations.github.service.base import GitHubHTTPClient
 from openhands.integrations.service_types import Comment
